Doc_summarizer/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
beautifulsoup4
wikipedia
graphviz
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
wikipedia
Flask
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
flask-cors
Markdown
googlesearch-python
pycountry
langchain-core
langchain_community
langchain_ollama
PyMuPDF
